Implement transaction timeouts (#6490)
## What is the goal of this PR?

We implement transaction timeouts (outined in https://github.com/vaticle/typedb/issues/6487), to prevent users from accidentally leaving transactions open indefinitely and leaking memory as a result. The transaction timeout is configurable via the usual transaction/session options.

## What are the changes implemented in this PR?

* Add transaction timeout that starts from the second the transaction is opened
* automatically close (with error) the transaction when the timeout occurs

+++ rocks/RocksSession.java
@@ -207,7 +207,7 @@ public abstract class RocksSession implements TypeDB.Session {
             long lock = 0;
             if (type == Arguments.Transaction.Type.WRITE) {
                 try {
-                    int timeout = options.schemaLockTimeoutMillis();
+                    long timeout = options.schemaLockTimeoutMillis();
                     lock = database().schemaLock().tryReadLock(timeout, MILLISECONDS);
                     if (lock == 0) throw TypeDBException.of(DATA_ACQUIRE_LOCK_TIMEOUT);
                 } catch (InterruptedException e) {
